// tt/deu-proj/src/models/transports.ts
export enum TransportsTypes {
  bus = 'bus',
  uBahn = 'ubahn',
  sBahn = 'sbahn',
  tram = 'tram',
  regionTraffic = 're',
  longDistanceTraffic = 'long-distance-traffic',
  walk = 'Walk',
  intercity = 'ec_ic',
  intercityExpress = 'ice',
  interRegio = 'ir',
  taxi = 'taxi',
  ferry = 'faehre',
  nordWestBahn = 'nwb',
  euroCity = 'ec'
}